Reorganized Modal subdir - Having separate Confirm and Prompt subdirectories cleans things up
