+++ packages/components/src/Modal/Prompt/index.tsx
@@ -24,6 +24,4 @@
 
  */
 
-export * from './Dialog'
-export * from './DialogManager'
-export * from './DialogSurface'
+export * from './Prompt'
